% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.binomialRF.R
\name{cv.binomialRF}
\alias{cv.binomialRF}
\title{random forest feature selection based on binomial exact test}
\usage{
cv.binomialRF(X, y, cvFolds = 5, fdr.threshold = 0.05,
  fdr.method = "BH", ntrees = 1000, percent_features = 0.5,
  keep.rf = F)
}
\arguments{
\item{X}{design matrix}

\item{y}{class label}

\item{cvFolds}{how many times should we perform cross-validation}

\item{fdr.threshold}{fdr.threshold for determining which set of features are significant}

\item{fdr.method}{how should we adjust for multiple comparisons (i.e., \code{p.adjust.methods} =c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY","fdr", "none"))}

\item{ntrees}{how many trees should be used to grow the \code{randomForest}? (Defaults to 5000)}

\item{percent_features}{what percentage of L do we subsample at each tree? Should be a proportion between (0,1)}

\item{keep.rf}{should we keep the randomForest object?}
}
\value{
a data.frame with 4 columns: Feature Name, cross-validated average for Frequency Selected, CV Median (Probability of Selecting it randomly), CV Median(Adjusted P-value based on \code{fdr.method}), and averaged number of times selected as signficant.
}
\description{
\code{cv.binomialRF} is the cross-validated form of the \code{binomialRF}, where K-fold crossvalidation is conducted to assess the feature's significance. Using the \code{cvFolds}=K parameter, will result in a K-fold cross-validation where the data is 'chunked' into K-equally sized groups and then the averaged result is returned.
}
\references{
Zaim, SZ. binomialRF: A novel randomForest feature selection algorithm. ArXiv, 2019.
}
