% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialRF_ensemble.R
\name{.binomialRF_ensemble}
\alias{.binomialRF_ensemble}
\title{a binomialRF_ensemble visualization and analysis for feature selection}
\usage{
.binomialRF_ensemble(candidateModels, X, y, ntrees = 2000, PLOT = F)
}
\arguments{
\item{candidateModels}{takes as input a list of candidate models by specifying which predictors you want to consider for each model as a string of vector names}

\item{X}{the design matrix X of predictors.}

\item{y}{the outcome y as a factor variable.}

\item{ntrees}{how many trees should be used to grow the \code{randomForest}? (Defaults to 5000)}

\item{PLOT}{should a graphical summary be released (boolean)}
}
\value{
a ggplot2 object comparison of different models
}
\description{
\code{binomialRF_ensemble} experimental section on plotting feature selection
}
\examples{
#' set.seed(324)
require(randomForest)
require(ggplot2)

###############################
### Generate simulation data
###############################

X = matrix(rnorm(1000), ncol=10)
trueBeta= c(rep(10,5), rep(0,5))
z = 1 + X \%*\% trueBeta
pr = 1/(1+exp(-z))
y = rbinom(100,1,pr)

###############################
### Run ensemble
###############################

candidateModels <- list(
      m1=c('X1','X4','X6','X9'),
      m2=c(paste('X',c(1,2,3,4,6,9),sep='')),
      m3=c(paste("X", 3:10,sep='')),
      m4= c(paste("X", 5:9,sep='')),
      m5= c(paste("X", c(4,5,8,10),sep='')),
      m6= c(paste("X",1:10,sep=''))
      )

}
