% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_correlbinom.R
\name{.fast_correlbinom}
\alias{.fast_correlbinom}
\title{a parallelized version of correlbinom}
\usage{
.fast_correlbinom(rho, successprob, trials, precision = 1024,
  model = "kuk")
}
\arguments{
\item{rho}{what percentage of data are you subsampling}

\item{successprob}{calculateBinomialP output}

\item{trials}{ntrees}

\item{precision}{how many bits in the pmf}

\item{model}{which model to use}
}
\value{
a ggplot2 object comparison of different models
}
\description{
\code{.fast_correlbinom} experimental section on plotting feature selection
}
