% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateCandidateModels.R
\name{evaluateCandidateModels}
\alias{evaluateCandidateModels}
\title{a binomialRf visualization for feature selection}
\usage{
evaluateCandidateModels(candidateModels, X, y, ntrees = 20000,
  percent_features = 0.2)
}
\arguments{
\item{candidateModels}{takes as input a list of candidate models by specifying which predictors you want to consider for each model as a string of vector names}

\item{X}{the design matrix X of predictors.}

\item{y}{the outcome y as a factor variable.}
}
\value{
a ggplot2 object comparison of different models
}
\description{
\code{evaluateCandidateModels} experimental section on plotting feature selection
}
\examples{
candidateModels <- list(
      m1=c('X1','X4','X6','X9'),
      m2=c(paste('X',c(1,2,3,4,6,9),sep='')),
      m3=c(paste("X", 3:10,sep='')),
      m4= c(paste("X", 5:9,sep='')),
      m5= c(paste("X", c(4,5,8,10),sep='')),
      m6= c(paste("X",1:10,sep=''))
      )

evaluateCandidateModels(candidateModels, X,y)
}
