% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneset_binomialRF.R
\name{geneset_binomialRF}
\alias{geneset_binomialRF}
\title{random forest feature selection based on binomial exact test}
\usage{
geneset_binomialRF(binomialRF_object, gene_ontology, cutoff = 0.2)
}
\arguments{
\item{binomialRF_object}{the binomialRF object output}

\item{gene_ontology}{a two- or three-column representation of a gene ontology with gene and geneset names}

\item{cutoff}{a real-valued number between 0 and 1, used as a p-value threshold}
}
\value{
a data.frame with 4 columns: Geneset Name, P-value, Adjusted P-value based on \code{fdr.method}
}
\description{
\code{binomialRF} is the R implementation of the feature selection algorithm by (Zaim 2019)
}
\references{
Zaim, SZ; Kenost, C.; Lussier, YA; Zhang, HH. binomialRF: Scalable Feature Selection and Screening for Random Forests to Identify Biomarkers and Their Interactions, bioRxiv, 2019.
}
