% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k.binomialRF.R
\name{k.binomialRF}
\alias{k.binomialRF}
\title{random forest feature selection based on binomial exact test}
\usage{
k.binomialRF(X, y, fdr.threshold = 1e-04, fdr.method = "BH",
  ntrees = 500, percent_features = floor(sqrt(ncol(X))), keep.rf = F,
  K = 2)
}
\arguments{
\item{X}{design matrix}

\item{y}{class label}

\item{fdr.threshold}{fdr.threshold for determining which set of features are significant}

\item{fdr.method}{how should we adjust for multiple comparisons (i.e., \code{p.adjust.methods} =c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY","fdr", "none"))}

\item{ntrees}{how many trees should be used to grow the \code{randomForest}? (Defaults to 5000)}

\item{percent_features}{what percentage of L do we subsample at each tree? Should be a proportion between (0,1)}

\item{keep.rf}{should we keep the randomForest object?}

\item{k}{how deep should the interactions be}
}
\value{
a data.frame with 4 columns: Feature Name, Frequency Selected, Probability of Selecting it randomly, Adjusted P-value based on \code{fdr.method}
}
\description{
\code{binomialRF} is the R implementation of the feature selection algorithm by (Zaim 2019). \code{k.binomialRF} extends the \code{binomialRF} algorithm by searching for k-way interactions.
}
\references{
Zaim, SZ. binomialRF: A novel randomForest feature selection algorithm. ArXiv, 2019.
}
